requirements.txt
flask
flask_cors
pandas
pillow
qrcode
openpyxl